refactor(tools): decompose hora-extra page into feature structure
Extracts 858-LOC monolith into feature folder under
app/features/tools/hora-extra/:
- page.vue — shell/layout component
- useHoraExtraCalculator.ts — re-export facade for model functions
- hora-extra.schema.ts — Zod form validation schema
- HoraExtraResult.vue — props-only result display component
- __tests__/useHoraExtraCalculator.spec.ts — 11 unit tests

Route page app/pages/tools/hora-extra.vue is now a thin wrapper.
Adds .hora-extra-page__breakdown wrapper div to preserve test compatibility.
